Moved common responses to common.js, made controllers for contacts and users
